Fix auth.js tests
This sets NODE_ENV to 'test' unconditionally on execution of the test.
As long as we'll use these environments, we'll have to make sure that
no time gets wasted looking for env differences.

In fact, test execution is itself a guarantee that .. we're in test. That's
why the environment variable is dangerous and bogus.