Fixes fluky tests: Extra AssertionError in socket.test.js
(With Thiebaud)

Our socket.test.js uses a socket.io library that reconnects disconnected
connections after a timeout. Our test tries to prevent this reconnection
by disconnecting and deleting the object manually, but js timers persist and
the object then tries to reconnect after 500ms. When reconnection fails,
the handler is called completely out of sequence.

We're adding an option here to prevent reconnects in general.