Refactor of Manager.js
Cleans up data modelling; as a result, a lot of code gets quite simpler.
Also changes strategy for detection of last socket disconnect in a
namespace: We now don't rely on socket-io anymore. The structures we
were using were private and not updated the way they should have been -
and the code is a mess. This will be easier to port to the next version.