Removes a broadly scoped test
The test says that it checks if Host is forwarded into the path properly. This
is also covered by the tests for the route implemented in this server.

Also, after fixing the error handling facility (in several stages), it turns
out that the route originally used for the test did not return a 200, but a
500 instead - because it tried to access an attachment that wasn't there.
Instead of rewriting a test that shows nothing, I decided to take the blue pill
and forget about it entirely.