Also verify expiry when accessing files
Refactors the code to also use MethodContext here; in there, we centralise
the checking code and call it from both paths. This ensures that no matter
how the access is loaded, expiry gets checked.

Added @flow to encryption.js - this makes it easier to track; also removed
the use of domain objects from encryption.js: It now deals just with strings.
This might seem more cumbersome, but is cleaner and easier to test.

Adds retrieveAccessById to MethodContext; this allows us to reuse the
check code. Unit tested.