Fixes 'fields' deprecation warning
We've cleaned our code already, using projection where we would use
fields before. There was an oversight (BUG!!) in the conversion methods
where we would still convert 'fields', setting it to undefined, thereby
triggering the depreaction warning. This fixes that piece of code,
renders usage of 'fields' to be a hard error and corrects the
FindOptions type signature.